Move the post index away from Vue. There won't be any dynamic bits going on for now.
