+++ app/Http/routes.php
@@ -18,7 +18,7 @@ $router->get('/admin/login', 'Admin\AuthController::loginForm');
 
 $router
     ->group('/api', function ($route) use ($apiCsrfMiddleware) {
-        $route->get('posts', 'API\PostController::index');
+        //
     })->middleware($authMiddleware);
 
 /**
