+++ src/sass/_style.scss
@@ -4,3 +4,4 @@
 @import "style/sidebar";
 @import "style/content";
 @import "style/form";
+@import "style/posts";
