+++ src/views/index.js
@@ -2,4 +2,3 @@
  * Import ALL THE VIEWS!
  */
 import login from './login';
-import postIndex from './post-index';
