--- src/views/post-index/style.scss
@@ -1,9 +1,0 @@
-@import '../../sass/organise';
-
-.call-to-action {
-    text-align: center;
-
-    @media screen and (min-width: 768px) {
-        padding-top: 2em;
-    }
-}
